==> Code/Regulator_contribution_Binsfeld25/workflow/01-data_exploration.Rmd <==
---
title: "01-data_exploration"
date: "Compiled at `r format(Sys.time(), '%Y-%m-%d %H:%M:%S', tz = 'UTC')` UTC"
output: github_document
params:
  name: "01-data_exploration" # change if you rename file
---

```{r here, message=FALSE}
here::i_am(paste0(params$name, ".Rmd"), uuid = "85d977d4-6a20-49dd-aa45-5d1eae842469")
```

The purpose of this document is to import CPI scores from each genetic background. Filter experiments and normalize the data.

```{r packages}
library(tidyverse)
```

```{r directories}
# create or *empty* the target directory, used to write this file's data: 
projthis::proj_create_dir_target(params$name, clean = TRUE)

# function to get path to target directory: path_target("sample.csv")
path_target <- projthis::proj_path_target(params$name)

# function to get path to previous data: path_source("00-import", "sample.csv")
path_source <- projthis::proj_path_source(params$name)
```

# Read files
  
First I will read each file for each genotype individually.

```{r read_scores}
# Wild-type.
wt_scores <- read.csv(path_source("00-import", "Scores_WT.txt"), sep='\t')
colnames(wt_scores) <- paste0("WT_", colnames(wt_scores))

wt_scores <- wt_scores %>% 
  
  # Make chemicals a named column
  rownames_to_column("chem_conc")

# marA deletion
dmara_scores <- read.csv(path_source("00-import","Scores_dmarA.txt"), sep='\t') %>% 
  rownames_to_column("chem_conc") 

# rob deletion
drob_scores <- read.csv(path_source("00-import", "Scores_drob.txt"), sep='\t') %>% 
  rownames_to_column("chem_conc") 

# soxS deletion 
dsoxs_scores <- read.csv(path_source("00-import", "Scores_dsoxS.txt"), sep='\t') %>% 
  rownames_to_column("chem_conc")
```


# Data Summaries  
  
## Wildtype

We start to get an idea of what our data looks like if we just look at the WildType

```{r brief_summary}
summary_str <- function(df){
  
  # Number of chemicals
  n_chems <- df %>% 
  mutate(chem_name = str_sub(chem_conc, start = 1, end = -2)) %>% 
  select(chem_name) %>% 
  n_distinct()
  
  # Number of concentrations
  n_concentrations <- df %>% 
    mutate(concentration = str_sub(chem_conc, start = -1)) %>% 
    select(concentration) %>% 
    n_distinct()
  
  # Number of total strains
  n_strains <- df %>% 
    select(-c(chem_conc)) %>% 
    ncol
  
  # Number of experiments
  n_experiments <- n_strains * n_chems * n_concentrations
  
  # Number of unique genes
  unique_genes <- unique(str_remove(colnames(df %>% select(-c(chem_conc))), "_\\d"))
  n_unique_genes <- length(unique_genes)

  
  sumstr <- paste("In this experiment there are\n\t", 
        n_chems, "chemicals tested at\n\t", 
        n_concentrations, "concentrations screened against\n\t",
        n_strains, "strains consisting of \n\t",
        n_unique_genes, "unique genes for which there are two replicates. This means a total of\n\t",
        n_experiments, "different assays were conducted")
  
  return(sumstr)
}
```
  

WT info  
  

```{r WT_summary}
cat(summary_str(wt_scores))
```
  
We can also take a look at the table itself

```{r wt_view}
head(wt_scores)
```



Let's take a look a the other experiments
  
## Pertubation Exploration  


dmarA summary
```{r mara_summary}

cat(summary_str(dmara_scores))

```
  

drob summary
```{r rob_summary}

cat(summary_str(drob_scores))

```
  
dsoxs summary
```{r soxs_summary}

cat(summary_str(dsoxs_scores))

```

  
# Failed experiments  
  
As expected, there will be some compound-concentration combinations where the rows are all NAs. This is mainly due to a high concentration of antibiotics. Let's try to detect those.

```{r failed_experiments}
number_nas <- function(df){
  
  # Get number of NAs per chem-conc combination
  na_info <- df %>% 
  
    # Transpose chemicals to columns
    column_to_rownames("chem_conc") %>% 
    t() %>% 
    as.data.frame() %>% 
    
    # Count number of NAs by column
    summarise_all(~ sum(is.na(.))) %>% 
    t() %>% 
    as.data.frame() %>% 
    
    # Format for joining
    rownames_to_column("chem_conc") %>% 
    rename("n_NAs" = "V1")
  
  
  df_with_na <- df %>% 
    left_join(na_info, by="chem_conc")
  
  return(df_with_na)
  
}

wt_scores <- number_nas(wt_scores)
dmara_scores <- number_nas(dmara_scores)
drob_scores <- number_nas(drob_scores)
dsoxs_scores <- number_nas(dsoxs_scores)

```

  
Let's see how many compounds failed completely, and whether they consistenly fail across all genotypes.
  
```{r failed_experiment_intersection}
N_STRAINS <- 16

failed_list <- list("WT" = wt_scores %>% dplyr::filter(n_NAs == N_STRAINS) %>% select(chem_conc) %>% unlist(),
                    "dmarA" = dmara_scores %>% dplyr::filter(n_NAs == N_STRAINS) %>% select(chem_conc) %>% unlist(),
                    "drob" = drob_scores %>% dplyr::filter(n_NAs == N_STRAINS) %>% select(chem_conc) %>% unlist(),
                    "dsoxs" = dsoxs_scores %>% dplyr::filter(n_NAs == N_STRAINS) %>% select(chem_conc) %>% unlist())

UpSetR::upset(UpSetR::fromList(failed_list))
```
  
The majority of failed experiments are failing in at least three of the screens.
   
# Partial Failure  
  
In the previous section, we looked at compound-concentration combinations that fail across all strains. Now lets look at combinations that only partially fail.  
  
```{r partial_failure}

partial_failure_counts <- function(df, genotype_str){
  
  df %>% 
  count(n_NAs) %>% 
  dplyr::filter(n_NAs > 0) %>% 
  mutate(genotype = genotype_str)
  
}


partial_failure_long <- bind_rows(partial_failure_counts(wt_scores, "WT"),
                                  partial_failure_counts(dmara_scores, "dmarA"),
                                  partial_failure_counts(drob_scores, "drob"),
                                  partial_failure_counts(dsoxs_scores, "dsoxS"))



partial_failure_long %>% 
  pivot_wider(id_cols = n_NAs, names_from = genotype, values_from = n, values_fill = 0) %>% 
  dplyr::arrange(n_NAs)
```
  

```{r joining_experiments}

complete_data <- wt_scores %>% 
  select(-n_NAs) %>% 
  left_join(dmara_scores %>% select(-n_NAs), by="chem_conc") %>%
  left_join(drob_scores %>% select(-n_NAs), by="chem_conc") %>% 
  left_join(dsoxs_scores %>% select(-n_NAs), by="chem_conc")


total_rows <- nrow(complete_data)
complete_failure_removal <- nrow(complete_data %>% dplyr::filter(!chem_conc %in% unlist(failed_list)))


complete_data_nacount <- number_nas(complete_data)
any_failure_removal <- nrow(complete_data_nacount %>% dplyr::filter(n_NAs == 0))


summary_removal <- paste("There are", total_rows, "total chemical - concentration combinations\n\t", 
      total_rows-complete_failure_removal, "are removed due to complete failure (", complete_failure_removal, "remain)\n\t",
      total_rows-any_failure_removal, "are removed due to partial or complete failure (", any_failure_removal, "remain)")

cat(summary_removal)
```
  
    
Remove chemical-concentration combinations that cause *any* measurement to fail.

```{r only_complete}
complete_scores <- complete_data_nacount %>% 
  dplyr::filter(n_NAs == 0) %>% 
  select(-n_NAs)
```


```{r score_longform}
complete_scores.long <- complete_scores %>% 
  pivot_longer(cols=c(-chem_conc), names_to = "genotype_promoter_rep") %>% 
  separate(genotype_promoter_rep, into = c("genotype", "promoter", "replicate"), remove = FALSE, sep="_") %>% 
  
  mutate(chem_name = str_sub(chem_conc, start = 1, end = -2),
         concentration = str_sub(chem_conc, start = -1),
         promoter_rep = paste(promoter, replicate, sep="_"))
```
  
# Score distribution  

  
## Scores by Promoter and Genotype  

```{r score_distrib, out.width='100%', out.height='100%'}
g <- complete_scores.long %>% 
  mutate(genotype_rep = paste(genotype, replicate, sep="_"),
         genotype_rep = factor(genotype_rep, levels = c("dsoxS_2", "dsoxS_1", 
                                                        "drob_2", "drob_1",
                                                        "dmarA_2", "dmarA_1",
                                                        "WT_2", "WT_1")),
         
         promoter = case_when(promoter == "EVC" ~ "EVC",
                              promoter == "pacrAB" ~ "acrABp",
                              promoter == "pmicF" ~ "micFp",
                              promoter == "pmarRAB" ~ "marRABp",
                              promoter == "pompF" ~ "ompFp",
                              promoter == "psoxS" ~ "soxSp",
                              promoter == "prob" ~ "robp",
                              promoter == "ptolC" ~ "tolCp"),
         
         promoter = factor(promoter, levels = c("EVC", "acrABp", "marRABp", "micFp",
                                                "ompFp", "robp", "soxSp", "tolCp"))) %>% 
  
  ggplot(aes(x=value, y=genotype_rep, fill=genotype)) +
  geom_violin() +
  geom_boxplot(width=0.2, fill="white", outlier.size=1) +
  
  facet_wrap(~promoter, ncol=4, scales = "free_x") +
  
  scale_y_discrete(labels = c("\u0394soxS_2", "\u0394soxS_1", 
                                "\u0394rob_2", "\u0394rob_1",
                                "\u0394marA_2", "\u0394marA_1",
                                "WT_2", "WT_1")) +
  
  scale_fill_manual(values = c("WT" = "#a2a39f", "dmarA" = "#ec9370", "drob" = "#48aa7d", "dsoxS" = "#556c9a")) +
  
  theme_bw() +
  theme(legend.position = "none",
        text = element_text(size=7),
        axis.text = element_text(size=6)) +
  labs(x="CPI Score", y="Strain - Replicate")
g
ggsave(path_target("score_distribution.png"), plot = g, dpi=300, width = 16, height = 8, units = "cm")

ggsave(path_target("score_distribution.pdf"), plot = g, dpi=300, width = 16, height = 8, units = "cm")
```  
    
  
A quick look at expression changes 


```{r}
chems_of_interest <- c("Paraquat_2")


chem_location <- complete_scores.long %>% 
  mutate(chem_conc = paste(chem_name, concentration, sep="_"),
         genotype_rep = paste(genotype, replicate, sep="_")) %>% 
  dplyr::filter(chem_conc %in% chems_of_interest & promoter == "pmicF")

g <- complete_scores.long %>% 
  mutate(genotype_rep = paste(genotype, replicate, sep="_")) %>% 
  dplyr::filter(promoter=="pmicF") %>% 
  
  ggplot(aes(y=value, x=genotype_rep)) +
  geom_violin() +
  geom_boxplot(width=0.1, fill="white") +
  
  geom_point(data=chem_location, aes(y=value, x=genotype_rep, color=chem_conc), size=3) +
  #geom_line(data=chem_location, aes(y=value, x=genotype_rep, color=chem_conc, group=chem_conc)) +
  
  scale_color_manual(values = c(Paraquat_2 = "red")) +
  
  theme_bw() +
  labs(y="Score", x="Genotype_Replicate", title = "micF Score distribution", color="Chemical Scores")

g
ggsave(filename = path_target("micf_score_paraquat.png"), plot=g, dpi=300)
```
```{r}

chems_of_interest <- c("Caffeine_1")


chem_location <- complete_scores.long %>% 
  mutate(chem_conc = paste(chem_name, concentration, sep="_"),
         genotype_rep = paste(genotype, replicate, sep="_")) %>% 
  dplyr::filter(chem_conc %in% chems_of_interest & promoter == "pmicF" & genotype == "WT")


g <- complete_scores.long %>% 
  mutate(genotype_rep = paste(genotype, replicate, sep="_")) %>% 
  dplyr::filter(promoter=="pmicF" & genotype == "WT") %>% 
  
  ggplot(aes(y=value, x=genotype_rep)) +
  geom_violin() +
  geom_boxplot(width=0.1, fill="white") +
  
  geom_point(data=chem_location, aes(y=value, x=genotype_rep, color=chem_conc), size=3) +
  #geom_line(data=chem_location, aes(y=value, x=genotype_rep, color=chem_conc, group=chem_conc)) +
  
  scale_color_manual(values = c(Caffeine_1 = "red")) +
  
  theme_bw() +
  labs(y="Expression", x="Genotype", title = "micF Expression", color="Chemical", subtitle = "Response to Caffeine")

g
ggsave(filename = path_target("micf_score_caffeine_wt.png"), plot=g, dpi=300)
```

      
Focusing on pacrAB, one can notice that the distribution of scores is much more compressed when marA is deleted! For all other backgrounds, it seems like its expression is unaffected

```{r parcAB_focus, out.width='100%'}
complete_scores.long %>% 
  mutate(genotype_rep = paste(genotype, replicate, sep="_")) %>% 
  dplyr::filter(promoter=="pacrAB") %>% 
  
  ggplot(aes(x=value, y=genotype_rep, fill=genotype)) +
  geom_violin() +
  geom_boxplot(width=0.1, fill="white") +

  
  theme_bw() +
  theme(legend.position = "none") +
  labs(x="Score", y="Genotype_Replicate", title = "pacrAB Score distribution")
```
  
In the case of micF, it seems like its expression is affected by the deletion of rob or marA.  
```{r micf_focus}
complete_scores.long %>% 
  mutate(genotype_rep = paste(genotype, replicate, sep="_")) %>% 
  dplyr::filter(promoter=="pmicF") %>% 
  
  ggplot(aes(x=value, y=genotype_rep, fill=genotype)) +
  geom_violin() +
  geom_boxplot(width=0.1, fill="white") +

  
  theme_bw() +
  theme(legend.position = "none") +
  labs(x="Score", y="Genotype_Replicate", title = "pmicF Score distribution")
```
  

```{r ompF_focus, out.width='100%'}

complete_scores.long %>% 
  mutate(genotype_rep = paste(genotype, replicate, sep="_")) %>% 
  dplyr::filter(promoter=="pompF") %>% 
  
  ggplot(aes(x=value, y=genotype_rep, fill=genotype)) +
  geom_violin() +
  geom_boxplot(width=0.1, fill="white") +

  
  theme_bw() +
  theme(legend.position = "none") +
  labs(x="Score", y="Genotype_Replicate", title = "pompF Score distribution")

```

```{r tolC_focus, out.width='100%'}
complete_scores.long %>% 
  mutate(genotype_rep = paste(genotype, replicate, sep="_")) %>% 
  dplyr::filter(promoter=="ptolC") %>% 
  
  ggplot(aes(x=value, y=genotype_rep, fill=genotype)) +
  geom_violin() +
  geom_boxplot(width=0.1, fill="white") +

  
  theme_bw() +
  theme(legend.position = "none") +
  labs(x="Score", y="Genotype_Replicate", title = "ptolC Score distribution")
```
  

```{r soxs_focus, out.width='100%'}
complete_scores.long %>% 
  mutate(genotype_rep = paste(genotype, replicate, sep="_")) %>% 
  dplyr::filter(promoter=="psoxS") %>% 
  
  ggplot(aes(x=value, y=genotype_rep, fill=genotype)) +
  geom_violin() +
  geom_boxplot(width=0.1, fill="white") +

  
  theme_bw() +
  theme(legend.position = "none") +
  labs(x="Score", y="Genotype_Replicate", title = "psoxS Score distribution")

```
  
```{r rob_focus, out.width='100%'}
complete_scores.long %>% 
  mutate(genotype_rep = paste(genotype, replicate, sep="_")) %>% 
  dplyr::filter(promoter=="prob") %>% 
  
  ggplot(aes(x=value, y=genotype_rep, fill=genotype)) +
  geom_violin() +
  geom_boxplot(width=0.1, fill="white") +

  
  theme_bw() +
  theme(legend.position = "none") +
  labs(x="Score", y="Genotype_Replicate", title = "prob Score distribution")

```
  
  
```{r mara_focus, out.width='100%'}
complete_scores.long %>% 
  mutate(genotype_rep = paste(genotype, replicate, sep="_")) %>% 
  dplyr::filter(promoter=="pmarRAB") %>% 
  
  ggplot(aes(x=value, y=genotype_rep, fill=genotype)) +
  geom_violin() +
  geom_boxplot(width=0.1, fill="white") +

  
  theme_bw() +
  theme(legend.position = "none") +
  labs(x="Score", y="Genotype_Replicate", title = "pmarRAB Score distribution")

```  
  
## Joint Distribution  
  
```{r rob_joint_distrib}

complete_scores.long %>% 
  mutate(replicate = paste("rep", replicate, sep="_")) %>% 
  dplyr::filter(promoter=="prob") %>% 
  
  ggplot(aes(x=value, y=replicate)) + 
  
  ggdist::stat_halfeye(
    adjust = .5, 
    width = .6, 
    .width = 0, 
    point_colour = NA,
    fill=alpha("#C5C5C5", 0.5)) +
   geom_boxplot(
    width = .25,
    outlier.colour = alpha("black", 0.5),
    outlier.size = 0.1
    
  )  +
  
  theme_classic() +
  labs(x="Score", y="Density", title = "prob joint score distribution")
```
  
Now soxS
  
```{r soxs_joint_distrib}

complete_scores.long %>% 
  mutate(replicate = paste("rep", replicate, sep="_")) %>% 
  dplyr::filter(promoter=="psoxS") %>% 
  
  ggplot(aes(x=value, y=replicate)) + 
  
  ggdist::stat_halfeye(
    adjust = .5, 
    width = .6, 
    .width = 0, 
    point_colour = NA,
    fill=alpha("#C5C5C5", 0.5)) +
   geom_boxplot(
    width = .25,
    outlier.colour = alpha("black", 0.5),
    outlier.size = 0.1
    
  )  +
  
  theme_classic() +
  labs(x="Score", y="Density", title = "psoxS joint score distribution")
```
  
Now marA
  
```{r mara_joint_distrib}

complete_scores.long %>% 
  mutate(replicate = paste("rep", replicate, sep="_")) %>% 
  dplyr::filter(promoter=="pmarRAB") %>% 
  
  ggplot(aes(x=value, y=replicate)) + 
  
  ggdist::stat_halfeye(
    adjust = .5, 
    width = .6, 
    .width = 0, 
    point_colour = NA,
    fill=alpha("#C5C5C5", 0.5)) +
   geom_boxplot(
    width = .25,
    outlier.colour = alpha("black", 0.5),
    outlier.size = 0.1
    
  )  +
  
  theme_classic() +
  labs(x="Score", y="Density", title = "pmarRAB joint score distribution")
```    
  
Now micF
  
```{r micf_joint_distrib}

complete_scores.long %>% 
  mutate(replicate = paste("rep", replicate, sep="_")) %>% 
  dplyr::filter(promoter=="pmicF") %>% 
  
  ggplot(aes(x=value, y=replicate)) + 
  
  ggdist::stat_halfeye(
    adjust = .5, 
    width = .6, 
    .width = 0, 
    point_colour = NA,
    fill=alpha("#C5C5C5", 0.5)) +
   geom_boxplot(
    width = .25,
    outlier.colour = alpha("black", 0.5),
    outlier.size = 0.1
    
  )  +
  
  theme_classic() +
  labs(x="Score", y="Density", title = "pmicF joint score distribution")
```      
 
Now ompF
  
```{r ompF_joint_distrib}

complete_scores.long %>% 
  mutate(replicate = paste("rep", replicate, sep="_")) %>% 
  dplyr::filter(promoter=="pompF") %>% 
  
  ggplot(aes(x=value, y=replicate)) + 
  
  ggdist::stat_halfeye(
    adjust = .5, 
    width = .6, 
    .width = 0, 
    point_colour = NA,
    fill=alpha("#C5C5C5", 0.5)) +
   geom_boxplot(
    width = .25,
    outlier.colour = alpha("black", 0.5),
    outlier.size = 0.1
    
  )  +
  
  theme_classic() +
  labs(x="Score", y="Density", title = "pompF joint score distribution")
```      
  
Now tolC
  
```{r tolC_joint_distrib}

complete_scores.long %>% 
  mutate(replicate = paste("rep", replicate, sep="_")) %>% 
  dplyr::filter(promoter=="ptolC") %>% 
  
  ggplot(aes(x=value, y=replicate)) + 
  
  ggdist::stat_halfeye(
    adjust = .5, 
    width = .6, 
    .width = 0, 
    point_colour = NA,
    fill=alpha("#C5C5C5", 0.5)) +
   geom_boxplot(
    width = .25,
    outlier.colour = alpha("black", 0.5),
    outlier.size = 0.1
    
  )  +
  
  theme_classic() +
  labs(x="Score", y="Density", title = "ptolC joint score distribution")
```        
  
Now acrAB
```{r acrAB_joint_distrib}

complete_scores.long %>% 
  mutate(replicate = paste("rep", replicate, sep="_")) %>% 
  dplyr::filter(promoter=="pacrAB") %>% 
  
  ggplot(aes(x=value, y=replicate)) + 
  
  ggdist::stat_halfeye(
    adjust = .5, 
    width = .6, 
    .width = 0, 
    point_colour = NA,
    fill=alpha("#C5C5C5", 0.5)) +
   geom_boxplot(
    width = .25,
    outlier.colour = alpha("black", 0.5),
    outlier.size = 0.1
    
  )  +
  
  theme_classic() +
  labs(x="Score", y="Density", title = "pacrAB joint score distribution")
```          
  
Finally EVC  
```{r evc_joint_distrib}

complete_scores.long %>% 
  mutate(replicate = paste("rep", replicate, sep="_")) %>% 
  dplyr::filter(promoter=="EVC") %>% 
  
  ggplot(aes(x=value, y=replicate)) + 
  
  ggdist::stat_halfeye(
    adjust = .5, 
    width = .6, 
    .width = 0, 
    point_colour = NA,
    fill=alpha("#C5C5C5", 0.5)) +
   geom_boxplot(
    width = .25,
    outlier.colour = alpha("black", 0.5),
    outlier.size = 0.1
    
  )  +
  
  theme_classic() +
  labs(x="Score", y="Density", title = "EVC joint score distribution")
```          
  
  
I would summarise that the distribution of each gene is quite distinct from each other. However, replicates of the same gene seem to be pretty similar. No clear bimodal distributions.  

## Quantile normalization  
  
Quantile normalization to account for baseline changes in expression due to removal of a gene. 


```{r qnorm_pmicF}
gene_expression <- complete_scores.long %>% 
  # Get the genotype_replicate
  mutate(genotype_rep = paste(genotype, replicate, sep="_"),
         chem_conc = paste(chem_name, concentration, sep = ":")) %>% 
  
  # filter for gene of interest
  dplyr::filter(promoter=="pmicF") %>% 
  
  # Pivot to matrix
  pivot_wider(id_cols = chem_conc, names_from = genotype_rep, values_from = value) %>% 
  column_to_rownames("chem_conc")

qnormed_scores <- as.data.frame(preprocessCore::normalize.quantiles(as.matrix(gene_expression)))

row.names(qnormed_scores) <- row.names(gene_expression)
colnames(qnormed_scores) <- colnames(gene_expression)

qnormed_scores.long <- qnormed_scores %>% 
  # Long form
  rownames_to_column("chem_conc") %>% 
  pivot_longer(cols = -c(chem_conc), names_to = "genotype_rep") %>% 
  
  # Get all original data back
  separate(chem_conc, into = c("chem_name", "concentration"), sep=":", remove=FALSE) %>% 
  separate(genotype_rep, into = c("genotype", "replicate"), sep="_", remove = FALSE)




qnormed_scores.long %>% 
  
  ggplot(aes(y=value, x=genotype_rep, fill=genotype)) +
  geom_violin() +
  geom_boxplot(width=0.1, fill="white") +
  
  theme_bw() +
  theme(legend.position = "none") +
  labs(x="Score", y="Genotype_Replicate", title = "QNORM-Score distribution by promoter")
```
```{r qnrom_caff}
chems_of_interest <- c("Caffeine:1")


chem_location <- qnormed_scores.long %>% 
  dplyr::filter(chem_conc %in% chems_of_interest)

g <- qnormed_scores.long %>% 
  ggplot(aes(y=value, x=genotype_rep)) +
  geom_violin() +
  geom_boxplot(width=0.1, fill="white") +
  
  geom_point(data=chem_location, aes(y=value, x=genotype_rep, color=chem_conc), size=3) +
  
  scale_color_manual(values=c(`Caffeine:1` = "red")) +
  
  theme_bw() +
  labs(y="Score", x="Genotype_Replicate", title = "micF", subtitle = "Quantile normalised score distribution", color="Chemical Score")


g
ggsave(filename = path_target("micf_qnorm_caffeine.png"), plot=g, dpi=300)
```


```{r qq_raw}
qnorm_by_gene <- function(gene_name, df = complete_scores.long){
  
  gene_expression <- df %>% 
    
    # Get the genotype_replicate
    mutate(genotype_rep = paste(genotype, replicate, sep="_"),
         chem_conc = paste(chem_name, concentration, sep = ":")) %>% 
    
    # filter for gene of interest
    dplyr::filter(promoter==gene_name) %>% 
    
    # Pivot to matrix
    pivot_wider(id_cols = chem_conc, names_from = genotype_rep, values_from = value) %>% 
    column_to_rownames("chem_conc")
  
  # Qnormed scores
  qnormed_scores <- as.data.frame(preprocessCore::normalize.quantiles(as.matrix(gene_expression)))
  
  row.names(qnormed_scores) <- row.names(gene_expression)
  colnames(qnormed_scores) <- colnames(gene_expression)
  
  # Qnorm long format
  qnormed_scores.long <- qnormed_scores %>% 
    # Long form
    rownames_to_column("chem_conc") %>% 
    pivot_longer(cols = -c(chem_conc), names_to = "genotype_rep") %>% 
    
    # Get all original data back
    separate(chem_conc, into = c("chem_name", "concentration"), sep=":", remove=FALSE) %>% 
    separate(genotype_rep, into = c("genotype", "replicate"), sep="_", remove = FALSE) %>% 
    
    mutate(promoter = gene_name)
  
  return(qnormed_scores.long)
  
}

qnorm_all.long <- bind_rows(lapply(unique(complete_scores.long$promoter), qnorm_by_gene))

```



```{r qnorm_treatment}

g <- qnorm_all.long %>% 
  
  ggplot(aes(x=value, y=genotype_rep, fill=genotype)) +
  geom_violin() +
  geom_boxplot(width=0.1, fill="white") +
  
  facet_wrap(~promoter, ncol=4, scales = "free") +
  
  theme_bw() +
  theme(legend.position = "none") +
  labs(x="Score", y="Genotype_Replicate", title = "QNORM-Score distribution by promoter")


g
ggsave(filename=path_target("qnormed_scores.png"), plot=g, dpi=300)

```


  
## Clustering

```{r pivot_zscores}
# Rows: Chems, # Columns: promoter_rep
normalised_scores.pivot <- qnorm_all.long %>% 
  mutate(promoter_rep = paste(promoter, replicate, sep="_"),
         chem_conc_genotype = paste(chem_name, concentration, genotype, sep="_")) %>% 
  pivot_wider(id_cols = chem_conc_genotype, names_from = promoter_rep, values_from = value) %>% 
  column_to_rownames("chem_conc_genotype")
```

```{r qbreaks}
quantile_breaks <- function(xs, n = 10) {
  breaks <- quantile(as.matrix(xs), probs = seq(0, 1, length.out = n))
  breaks[!duplicated(breaks)]
}

mat_breaks <- quantile_breaks(normalised_scores.pivot, n = 10)
mat_breaks
```



```{r heatmap_colors}
pheatmap::pheatmap(normalised_scores.pivot, 
                   
                   # Clustering Params
                   clustering_distance_cols = "correlation",
                   clustering_distance_rows = "correlation",
                   clustering_method = "ward.D2",
                  
                   # Presentation Details
                   show_rownames = FALSE,
                   breaks = mat_breaks,
                   color=rev(RColorBrewer::brewer.pal(length(mat_breaks)-1, "RdBu")))
```

Well, I think the most obvious conclusion from this is that replicates cluster together. That's good!!
    
  
### Specific Compounds

I'm just generally curious to see how some compounds behave across genotypes

```{r micf_compounds_interest, out.width='100%'}

chems_of_interest <- c("Caffeine_1", "Salicylate_1")


chem_location <- qnorm_all.long %>% 
  mutate(chem_conc = paste(chem_name, concentration, sep="_"),
         genotype_rep = paste(genotype, replicate, sep="_")) %>% 
  dplyr::filter(chem_conc %in% chems_of_interest & promoter == "pmicF")

qnorm_all.long %>% 
  mutate(genotype_rep = paste(genotype, replicate, sep="_")) %>% 
  dplyr::filter(promoter=="pmicF") %>% 
  
  ggplot(aes(y=value, x=genotype_rep)) +
  geom_violin() +
  geom_boxplot(width=0.1, fill="white") +
  
  geom_point(data=chem_location, aes(y=value, x=genotype_rep, color=chem_conc)) +

  
  theme_bw() +
  labs(y="Score", x="Genotype_Replicate", title = "pmicF Score distribution", color="Chemical")
```
  
Now this is a super cool phenomenon! 
When rob is present then caffeine is always among the top hits. Once rob is deleted, Caffeine is longer among the top!!!
Sort of the inverse is happening with Salicylate!
Amazing honestly.  
  
What happens to these compounds in the promoters?

```{r mara_compounds_interest, out.width='100%'}

chems_of_interest <- c("Caffeine_1", "Salicylate_1")


chem_location <- qnorm_all.long %>% 
  mutate(chem_conc = paste(chem_name, concentration, sep="_"),
         genotype_rep = paste(genotype, replicate, sep="_")) %>% 
  dplyr::filter(chem_conc %in% chems_of_interest & promoter == "pmarRAB")

qnorm_all.long %>% 
  mutate(genotype_rep = paste(genotype, replicate, sep="_")) %>% 
  dplyr::filter(promoter=="pmarRAB") %>% 
  
  ggplot(aes(y=value, x=genotype_rep)) +
  geom_violin() +
  geom_boxplot(width=0.1, fill="white") +
  
  geom_point(data=chem_location, aes(y=value, x=genotype_rep, color=chem_conc)) +

  
  theme_bw() +
  labs(y="Score", x="Genotype_Replicate", title = "pmarRAB Score distribution", color="Chemical")

```

```{r rob_compounds_interest, out.width='100%'}

chems_of_interest <- c("Caffeine_1", "Salicylate_1")


chem_location <- qnorm_all.long %>% 
  mutate(chem_conc = paste(chem_name, concentration, sep="_"),
         genotype_rep = paste(genotype, replicate, sep="_")) %>% 
  dplyr::filter(chem_conc %in% chems_of_interest & promoter == "prob")

qnorm_all.long %>% 
  mutate(genotype_rep = paste(genotype, replicate, sep="_")) %>% 
  dplyr::filter(promoter=="prob") %>% 
  
  ggplot(aes(y=value, x=genotype_rep)) +
  geom_violin() +
  geom_boxplot(width=0.1, fill="white") +
  
  geom_point(data=chem_location, aes(y=value, x=genotype_rep, color=chem_conc)) +

  
  theme_bw() +
  labs(y="Score", x="Genotype_Replicate", title = "prob Score distribution", color="Chemical")

```
  
## Write Files

```{r write_output}
write_tsv(complete_scores.long,path_target("raw_scores_complete.tsv.gz"))
write_tsv(qnorm_all.long, path_target("qnormed_scores_complete.tsv.gz"))
```  
  
  
## Files written

These files have been written to the target directory, ```r paste0("data/", params$name)```:

```{r list-files-target}
projthis::proj_dir_info(path_target())
```

## Session Info. 
  
```{r}
sessionInfo()
```
